requirements.txt
PyQt6
ebooklib
Pillow
beautifulsoup4
htmlmin
cssmin
jsmin